backend/requirements.txt
asgiref==3.8.1
Django==5.2.2
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
psycopg2-binary>=2.9.0
djangorestframework-3.16.0
djangorestframework-simplejwt-5.5.0
celery==5.5.3 
django-redis-6.2.0
drf-spectacular-0.29.1
django-cors-headers==3.18.0
django-prometheus-2.4.0
requests-2.32.4
django-phonenumber-field-8.1.0
drf-yasg-1.21.10
coreschema-0.0.4